refactor(core): crunch some code, minus should be minus
This commit introduce a new era. I just realized, after a big project
using minus, that minus should do one thing, and only one. It is not
intended to be an API, nor a tennisman, nor a cooker. So, I removed the
API part. Also, I just fought with the new version of piping (and the
bugs of the old version), and I'm so tired of fighting with that
unmaintened repo that I realized too that it was almost useless
without the API part. So, bye piping. Keep the long way - LOL -

Minus is intended to be a *front* boilerplate, with isomorphic
abilities. So, it feel less wrong to me with a *reduced* server part.

Also, the "middlewares" folder in "server" folder has been deleted. A
bit less semantic, but a bit easier to read/edit the render file.

Startup time is now faster, and src LOC is now 294. Yeah!
Note that I've been forced to put some content in the `Home` component
due to a recent eslint rule suprise update.

Feel free to expose your opinion, and keep in mind that I will always
respect it.
