+++ bin/start.js
@@ -5,23 +5,8 @@ require('babel-polyfill')
 require('ignore-styles')
 
 const path = require('path')
-const piping = require('piping')
 const appModulePath = require('app-module-path')
-
-const config = require('../src/config').default
 
 appModulePath.addPath(path.resolve(__dirname, '../src'))
 
-const launcher = {
-  '--app': () => require('../src/server'),
-  '--api': () => require('../src/server/dev-api'),
-}
-
-const pipingOpts = {
-  hook: true,
-  ignore: /(actions|components|reducers|styles)\/.*/,
-}
-
-if (config.env !== 'development' || piping(pipingOpts)) {
-  launcher[process.argv[2]]()
-}
+require('../src/server')
