+++ src/server/index.js
@@ -3,7 +3,7 @@ import compression from 'compression'
 import bodyParser from 'body-parser'
 
 import config from 'config'
-import render from 'server/middlewares/render'
+import render from 'server/render'
 
 const server = express()
 
@@ -24,6 +24,6 @@ server.use(render)
 server.listen(config.port, 'localhost', err => {
   /* eslint-disable no-console */
   if (err) { return console.log(err) }
-  console.log(`[APP] listening at localhost:${config.port} in ${config.env} mode`)
+  console.log(`\n[APP] listening at localhost:${config.port} in ${config.env} mode`)
   /* eslint-enable no-console */
 })
